mypro1/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2==2.9.1
django-crispy-forms==1.12.0
python-dotenv==0.19.0
django-cors-headers==3.8.0
coverage==6.0
